+++ CHANGELOG.md
@@ -38,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added integration tests for MPT transactions `MPTokenAuthorize`, `MPTokenIssuanceCreate`, `MPTokenIssuanceDestroy` and `MPTokenIssuanceSet`
 - Added `RippleTimeToUnixSeconds` function
 - Added `GetAMMInfo` query for both RPC and WebSocket clients
-- Added integration test for amm transaction `AMMCreate`
+- Added integration test for amm transactions
 ### Changed
 
 #### Makefile